Doze Mode & App Standby Compatibility
Support Android 6.0 (Marshmallow) devices by adding a dialog that asks
the user to disable doze mode explicitly for RedAlert, so that rocket
alerts may be received in real time (through the secondary push
notification gateway).

GCM push notifications are sent as high-priority anyway, but we want the
secondary push gateway to work as well, just in case.