v1.0.8.6 - Revert Doze Compat (Google Play Policy)
Unfortunately, Google Play decided to suspend the app and that the only
way it will be reinstated is if the the
REQUEST_IGNORE_BATTERY_OPTIMIZATIONS permission is removed, even though
I explained that according to the docs I had every right to request it.
I decided to remove it so that I won't risk permanent deletion of the
app.

It seems that Google is extremely trigger-happy and emotionless about
suspending and permanently banning apps on Google Play.