Fix `RuntimeException` in Android 4.0 devices (Mi Band Integration)
Due to a bug in Android 4.0, apps cannot query for the
`BluetoothAdapter.getDefaultAdapter()` from an async (non-UI) thread.

The workaround is to cache the adapter on a UI thread, or call
`Looper.prepare()` on the async thread.

I chose to call `Looper.prepare()` because I don’t have an efficient
way to call `BluetoothAdapter.getDefaultAdapter()` from a
Looper-enabled thread when the application is invoked via a push
notification. I do not have access to an activity with a UI thread in
the push BroadcastReceiver.

http://stackoverflow.com/questions/5920578/bluetoothadapter-getdefault-t
hrowing-runtimeexception-while-not-in-activity